Confirm status auth status with backend before operation (that needed the perms)

+++ frontend/src/functions/HttpRequest.js
@@ -1,4 +1,5 @@
 import apiClient from '@/api';
+import { authStatus, isAuth } from './Authentications';
 
 export async function getCsrfToken() {
     var csrfResponse = await apiClient.get(`/activities/get-csrf-token/`);
@@ -10,6 +11,8 @@ export async function createPostRequest(path, data) {
      * Create post request with CSRF token in its headers.
      * Return HttpResponse from post request
      */
+    await authStatus();
+    if (!isAuth.value) return; 
     const csrfToken = await getCsrfToken();
     return await apiClient.post(path, data, {
         headers: { 'X-CSRFToken': csrfToken },
@@ -21,6 +24,8 @@ export async function createPutRequest(path, data) {
      * Create PUT request with CSRF token in its headers.
      * Return HttpResponse from PUT request
      */
+    await authStatus();
+    if (!isAuth.value) return; 
     const csrfToken = await getCsrfToken();
     return await apiClient.put(path, data, {
         headers: { 'X-CSRFToken': csrfToken },
@@ -32,6 +37,8 @@ export async function createDeleteRequest(path) {
      * Create DELETE request with CSRF token in its headers.
      * Return HttpResponse from DELETE request
      */
+    await authStatus();
+    if (!isAuth.value) return; 
     const csrfToken = await getCsrfToken();
     return await apiClient.delete(path, {
         headers: { 'X-CSRFToken': csrfToken },
